Renamed wir and dag post/preprocessor for improved clarity
