+++ mlinspect/backends/_sklearn_wir_processor.py
@@ -34,7 +34,7 @@ class SklearnWirPreprocessor:
         self.wir_node_to_sub_pipeline_start = {}
         self.wir_node_to_sub_pipeline_end = {}
 
-    def preprocess_wir(self, graph: networkx.DiGraph) -> networkx.DiGraph:
+    def process_wir(self, graph: networkx.DiGraph) -> networkx.DiGraph:
         """
         Preprocess scikit-learn pipeline operations to hide the special pipeline
         declaration style from other parts of the library
