+++ test/instrumentation/test_wir_to_dag_transformer.py
@@ -4,7 +4,7 @@ Tests whether the DAG extraction works
 import networkx
 from testfixtures import compare
 
-from mlinspect.backends._sklearn_wir_preprocessor import SklearnWirPreprocessor
+from mlinspect.backends._sklearn_wir_processor import SklearnWirPreprocessor
 from mlinspect.instrumentation._dag_node import CodeReference
 from mlinspect.instrumentation._wir_extractor import WirExtractor
 from mlinspect.instrumentation._wir_node import WirNode
@@ -35,7 +35,7 @@ def test_remove_all_non_operators_and_update_names():
     """
     Tests whether the WIR Extraction works for the adult_easy pipeline
     """
-    preprocessed_wir = SklearnWirPreprocessor().preprocess_wir(get_test_wir())
+    preprocessed_wir = SklearnWirPreprocessor().process_wir(get_test_wir())
     cleaned_wir = WirToDagTransformer().remove_all_nodes_but_calls_and_subscripts(preprocessed_wir)
     dag = WirToDagTransformer.remove_all_non_operators_and_update_names(cleaned_wir)
